update relative pathing for inputs; assume bf16 for gemm and swiglu

+++ stream/workload/computation/computation_node.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 from math import prod
 from typing import TypeAlias
 
-from xdsl.dialects.builtin import MemRefType, i8, i16, i32
+from xdsl.dialects.builtin import MemRefType, i8, i32, bf16, f32
 from xdsl.dialects.memref import AllocOp, SubviewOp
 from zigzag.datatypes import Constants, LayerDim, LayerOperand, MemoryOperand
 from zigzag.utils import hash_sha512
@@ -24,8 +24,8 @@ LOOP_RANGES_T: TypeAlias = dict[LayerDim, tuple[int, int]]
 
 PRECISION_TYPE_MAP = {
     8: i8,
-    16: i16,
-    32: i32,
+    16: bf16,
+    32: f32,
 }
